backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai-whisper
torch
transformers
gTTS
sounddevice
python-dotenv==1.0.0
python-multipart==0.0.6